fix: separação de modulos , camada de infra e  configuração de imports.
