+++ src/modules/appointments/services/CreateAppointmentService.ts
@@ -2,10 +2,9 @@ import { startOfHour } from 'date-fns';
 
 import { getCustomRepository } from 'typeorm';
 
-import Appointment from '../models/Appointment';
+import AppError from '@shared/errors/AppError';
+import Appointment from '../infra/typeorm/entities/Appointment';
 import AppointmentsRepository from '../repositories/AppointmentsRepository';
-
-import AppError from '../errors/AppError';
 
 interface Request {
   date: Date;
