+++ src/modules/users/infra/http/middlewares/ensureAuthenticatedMiddleware.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import { verify } from 'jsonwebtoken';
-import authConfig from '../config/authConfig';
+import authConfig from '@config/authConfig';
 
-import AppError from '../errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 interface RequestToken {
   iat: number;
