+++ src/modules/users/services/AuthenticateUserService.ts
@@ -1,11 +1,11 @@
 import { getRepository } from 'typeorm';
 import { compare } from 'bcryptjs';
 import { sign } from 'jsonwebtoken';
-import authConfig from '../config/authConfig';
+import authConfig from '@config/authConfig';
 
-import User from '../models/User';
+import User from '@modules/users/infra/typeorm/entities/User';
 
-import AppError from '../errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 interface Request {
   email: string;
