+++ src/shared/infra/http/server.ts
@@ -4,12 +4,11 @@ import cors from 'cors';
 import express, { Request, Response, NextFunction } from 'express';
 import 'express-async-errors';
 
+import uploadConfig from '@config/uploadConfig';
+import AppError from '@shared/errors/AppError';
 import routes from './routes';
 
-import uploadConfig from './config/uploadConfig';
-import AppError from './errors/AppError';
-
-import './database';
+import '@shared/infra/typeorm';
 
 const app = express();
